requirements.txt
flask==3.0.3
gunicorn==21.2.0
fonttools==4.53.1
numpy==1.26.4
shapely==2.0.4
scikit-image==0.22.0
scipy==1.11.4
networkx==3.2.1
imageio==2.35.1
tifffile==2024.8.30
pillow==10.4.0
svgwrite==1.4.3
pandas==2.2.2
openpyxl==3.1.5




Flask-Cors>=4,<5
